Enabling role based access and adding additional refactoring
