+++ life-manager-rest/src/main/java/backend/app/lifemanager/course/CourseResource.java
@@ -19,4 +19,9 @@ public class CourseResource {
 	public List<Course> getAllCourses(@PathVariable String username) {
 		return courseManagementService.findAll();
 	}
+
+	@GetMapping("/instructors/{username}/courses/{id}")
+	public Course getCourse(@PathVariable String username, @PathVariable long id) {
+		return courseManagementService.findById(id);
+	}
 }
