+++ life-manager-rest/src/main/java/backend/app/lifemanager/course/CoursesHardcodedService.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import java.util.ArrayList;
 import java.util.List;
 
+//Todo: replace this with proper DB interaction implementation
 @Service
 public class CoursesHardcodedService {
 
@@ -21,5 +22,27 @@ public class CoursesHardcodedService {
 
 	public List<Course> findAll() {
 		return courses;
+	}
+
+	public Course findById(long id) {
+		for (Course course: courses) {
+			if (course.getId() == id) {
+				return course;
+			}
+		}
+		return null;
+	}
+
+	public Course deleteById(long id) {
+		Course course = findById(id);
+
+		if (course == null)
+			return null;
+
+		if (courses.remove(course)) {
+			return course;
+		}
+
+		return null;
 	}
 }
